requirements.txt
python-telegram-bot==22.3
aiohttp==3.9.5
